# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  LoginCredentials:
    properties:
      email:
        description: user's email address
        type: string
        x-go-name: Email
      password:
        description: user's password
        type: string
        x-go-name: PassWord
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/session
  LoginResponse:
    description: LoginResponse defines the struct for API login
    properties:
      message:
        description: Contains user information
        properties:
          email:
            description: the user's email address
            type: string
            x-go-name: Email
          firstname:
            description: the user's first name
            type: string
            x-go-name: FirstName
          id:
            description: the user's id
            type: string
            x-go-name: ID
          lastname:
            description: the user's last name
            type: string
            x-go-name: LastName
          password:
            description: the user's hashed password
            type: string
            x-go-name: PassWord
        type: object
        x-go-name: Message
      status:
        description: shows if process was successful or not
        type: boolean
        x-go-name: Status
      token:
        description: A unique number that defines authenticates the users transactions
        type: string
        x-go-name: Token
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/session
  MyStdResp:
    properties:
      message:
        description: delete message / error message when delete is unsuccessful
        type: string
        x-go-name: Message
      status:
        description: successful / not successful
        type: boolean
        x-go-name: Status
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/controllers
  PhoneBookContact:
    description: phonebook struct
    properties:
      firstname:
        type: string
        x-go-name: FirstName
      id:
        type: string
        x-go-name: ID
      lastname:
        type: string
        x-go-name: LastName
      phone_number:
        type: string
        x-go-name: PhoneNumber
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/models
  Tk:
    properties:
      token:
        type: string
        x-go-name: Token
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/session
  UserCredentials:
    description: For registering new users
    properties:
      email:
        type: string
        x-go-name: Email
      firstname:
        type: string
        x-go-name: FirstName
      id:
        type: string
        x-go-name: ID
      lastname:
        type: string
        x-go-name: LastName
      password:
        type: string
        x-go-name: PassWord
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/models
  deleteRequest:
    properties:
      id:
        type: string
        x-go-name: ID
      token:
        type: string
        x-go-name: Token
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/controllers
  updateRequest:
    properties:
      details:
        $ref: '#/definitions/PhoneBookContact'
      token:
        type: string
        x-go-name: Token
    type: object
    x-go-package: github.com/IamNator/mysql-golang-web/controllers
info:
  contact:
    email: natverior1@gmail.com
  description: Documentation for mysql-golang-web API
  title: mysql-golang-web API
  version: 1.0.0
paths:
  /api/contacts:
    delete:
      description: delete contact from phone book
      operationId: delete
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/deleteRequest'
      responses:
        "200":
          $ref: '#/responses/deleteResponse'
        "400":
          $ref: '#/responses/deleteUnauthorized'
        "500":
          $ref: '#/responses/deleteInternalError'
      tags:
      - controllers
    get:
      description: returns all phonebook contacts
      operationId: fetch
      parameters:
      - description: token
        in: header
        name: Authorization
        type: string
      responses:
        "200":
          $ref: '#/responses/fetchResponse'
        "401":
          $ref: '#/responses/fetchUnauthorized'
        "500":
          $ref: '#/responses/fetchInternalError'
      tags:
      - controllers
    put:
      description: adds new contacts to a phoneBook
      operationId: update
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/updateRequest'
      responses:
        "200":
          $ref: '#/responses/updateResponse'
        "401":
          $ref: '#/responses/updateUnauthorized'
        "409":
          $ref: '#/responses/updateConflictError'
        "500":
          $ref: '#/responses/updateInternalError'
      tags:
      - controllers
  /user/login:
    post:
      description: Returns a session token
      operationId: login
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/LoginCredentials'
      responses:
        "200":
          $ref: '#/responses/loginResponse'
        "400":
          $ref: '#/responses/loginBadRequest'
        "404":
          $ref: '#/responses/loginIncorrectPassword'
          description: '&& loginNotFound'
      tags:
      - session
  /user/logout:
    post:
      description: logs the user out
      operationId: logout
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Tk'
      responses:
        "200":
          $ref: '#/responses/logoutResponse'
        "400":
          $ref: '#/responses/logoutBadRequest'
        "500":
          $ref: '#/responses/logoutInternalError'
      tags:
      - session
  /user/register:
    post:
      operationId: register
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/UserCredentials'
      responses:
        "201":
          $ref: '#/responses/registerResponse'
        "302":
          $ref: '#/responses/registerUserExist'
        "500":
          $ref: '#/responses/registerInternalError'
      tags:
      - session
produces:
- application/json
responses:
  deleteInternalError:
    description: unable to respond to request
    schema:
      $ref: '#/definitions/MyStdResp'
  deleteResponse:
    description: delete successful
    schema:
      $ref: '#/definitions/MyStdResp'
  deleteUnauthorized:
    description: token not valid, login to get a new one
    schema:
      $ref: '#/definitions/MyStdResp'
  fetchInternalError:
    description: unable to respond to request
    schema:
      $ref: '#/definitions/MyStdResp'
  fetchResponse:
    description: ""
    headers:
      Body:
        description: in body
    schema:
      properties:
        message:
          items:
            $ref: '#/definitions/PhoneBookContact'
          type: array
          x-go-name: Message
        status:
          type: boolean
          x-go-name: Status
      type: object
  fetchUnauthorized:
    description: token not valid, login to get a new one
    schema:
      $ref: '#/definitions/MyStdResp'
  loginBadRequest:
    description: request does not conform to login api
    schema:
      $ref: '#/definitions/MyStdResp'
  loginIncorrectPassword:
    description: Incorrect password
    schema:
      $ref: '#/definitions/MyStdResp'
  loginNotFound:
    description: Non registered user tries to login
    schema:
      $ref: '#/definitions/MyStdResp'
  loginResponse:
    description: Successful login request
    schema:
      $ref: '#/definitions/LoginResponse'
  logoutBadRequest:
    description: unable to process request
    schema:
      $ref: '#/definitions/MyStdResp'
  logoutInternalError:
    description: Unable to respond
    schema:
      $ref: '#/definitions/MyStdResp'
  logoutResponse:
    description: ""
    schema:
      $ref: '#/definitions/MyStdResp'
  registerInternalError:
    description: returns when there is an internal server error (likely unable access
      database)
    schema:
      $ref: '#/definitions/MyStdResp'
  registerResponse:
    description: user successfully created
    schema:
      $ref: '#/definitions/MyStdResp'
  registerUserExist:
    description: when a user already exists
    schema:
      $ref: '#/definitions/MyStdResp'
  updateConflictError:
    description: phone number already exists
    schema:
      $ref: '#/definitions/MyStdResp'
  updateInternalError:
    description: unable to respond to request
    schema:
      $ref: '#/definitions/MyStdResp'
  updateResponse:
    description: update successful
    schema:
      properties:
        message:
          type: string
          x-go-name: Message
        status:
          type: boolean
          x-go-name: Status
      type: object
  updateUnauthorized:
    description: token not valid, login to get a new one
    schema:
      $ref: '#/definitions/MyStdResp'
schemes:
- http
swagger: "2.0"